Add group name to page title setting
The code introduces a new setting in the osu_groups_basic_group module that allows the automatic addition of the group name to the page title. A checkbox has been added to the group settings form to enable or disable this feature. The setting is also stored and retrieved using the Drupal configuration service. An update hook is included to set this new group global setting for auto node title by default.